bodyPartView refactored
bodyPart to views(folder name change)

methods moved from utils if only usable by one class

saving bodyPart and retrieving it from DB

workLogs loaded for only chosen bodyPart
