--- lib/bodyPart/bodyPartInterface.dart
@@ -1,12 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:workout_log/entity/workLog.dart';
-
-abstract class BodyPartInterface {
-  static List<Widget> wList = List();
-
-  updateWorkLogToDB(WorkLog workLog);
-
-  updateState();
-
-  void updateWorkLogFromDB() {}
-}
