+++ lib/entity/bodyPart.dart
@@ -1,4 +1,5 @@
 
 enum BodyPart {
-  CHEST, BACK, ARM, LEG
+  CHEST, BACK, ARM, LEG, UNDEFINED
+
 }
