+++ lib/entity/exercise.dart
@@ -1,6 +1,7 @@
 import 'package:json_annotation/json_annotation.dart';
 import 'package:uuid/uuid.dart';
 import 'package:workout_log/entity/bodyPart.dart';
+import 'package:workout_log/util/util.dart';
 
 part 'exercise.g.dart';
 
@@ -32,22 +33,26 @@ class Exercise {
 
   //  needed for SQLite
   factory Exercise.fromMap(Map<String, dynamic> json) {
+
+    BodyPart bp = Util.recreateBodyPart(json["bodyPart"]);
+
     Exercise e = Exercise(
       //  get from given map
       json["name"],
-      json["bodyPart"],
+      bp,
     );
-    //  id needed to be saved as it is in json,
-    //  otherwise id will again generated,
-    //  which will be incoherent with DB entry
-     e.id=json["id"];
+    ///  id needed to be saved as it is in json,
+    ///  otherwise id will again generated,
+    ///  which will be incoherent with DB entry
+    e.id = json["id"];
     return e;
   }
 
-  Map<String, dynamic> toMap() =>
-      {
+  Map<String, dynamic> toMap() => {
         "id": id,
         "name": name,
-        "bodyPart": bodyPart.toString(),
+        "bodyPart":
+            bodyPart.toString().substring(bodyPart.toString().indexOf('.') + 1),
       };
+
 }
