+++ pubspec.yaml
@@ -10,7 +10,7 @@ description: Save workouts logs !
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.2.0 <3.0.0"
 
 dependencies:
   flutter:
